Move api/utils/grpc to lib/utils/grpc
